Cleanup Browser Runtime Code
