+++ src/runtime/wasm/ecs/mod.rs
@@ -1,0 +1,5 @@
+
+mod info;
+mod query;
+mod resource;
+mod value;
